01-german-credit-mlops/requirements.txt
pandas==2.2.2
numpy==1.26.4

scikit-learn==1.4.2

matplotlib==3.8.4
seaborn==0.13.2

mlflow==2.12.1

shap==0.44.1

dvc==3.51.2

joblib==1.4.2
pyyaml==6.0.1


cloudpickle==3.0.0


numba==0.59.0
slicer==0.0.7
slicerutil==0.2.2

dvc-s3
dagshub
mlflow
python-dotenv==1.0.1

setuptools==69.5.1